refactor:dot 컴포넌트 분리
