+++ src/types/components/Dot.tsx
@@ -1,0 +1,7 @@
+import DotIcon from "@/assets/img/dot_icon.svg?react";
+
+const Dot = ({ margin = "0 4px 3px 4px" }: { margin?: string }) => (
+	<DotIcon style={{ margin }} />
+);
+
+export default Dot;
